feat: implement todo app with authentication, authorization, and CRUD operations
