+++ libs/todo-app/src/config.ts
@@ -1,0 +1,2 @@
+export const PORT = Number(process.env.PORT) || 3000;
+export const JWT_SECRET = process.env.JWT_SECRET || 'todo-app-dev-secret';
